Create power.sh
this is a script for displaying powerful information regarding
battery device information, mostly for laptops, however
other devices running off of power reserves could implement
this utility as well.
I'm unsure as to whether or not other *nix derivatives would
be able to support this exact implementation, if not, I'll
have to adapt it for that as well. Macintosh does not support
freedesktop, and most likely has its own power management utility,
and as such, this script is obsolete on that platform.